+++ src/SpriteSet.cpp
@@ -14,7 +14,8 @@ using namespace rgl;
 
 SpriteSet::SpriteSet(Material& in_material, int in_nvertex, double* in_vertex, int in_nsize, double* in_size,
                      int in_ignoreExtent, int count, Shape** in_shapelist, double* in_userMatrix,
-                     bool in_fixedSize, Scene *in_scene, double* in_adj,
+                     bool in_fixedSize, bool in_rotating,
+                     Scene *in_scene, double* in_adj,
                      int in_npos, int *in_pos, double in_offset)
  : Shape(in_material, in_ignoreExtent, SHAPE, true), 
   vertex(in_nvertex, in_vertex),
@@ -22,6 +23,7 @@ SpriteSet::SpriteSet(Material& in_material, int in_nvertex, double* in_vertex, i
    pos(in_npos, in_pos),
    offset(in_offset),
    fixedSize(in_fixedSize),
+   rotating(in_rotating),
    scene(in_scene)
 { 
   if (!count)
@@ -99,7 +101,7 @@ void SpriteSet::drawBegin(RenderContext* renderContext)
 
   m = Matrix4x4(renderContext->subscene->modelMatrix);
   
-  if (fixedSize) {
+  if (fixedSize && !rotating) {
     
     p = Matrix4x4(renderContext->subscene->projMatrix);
   
@@ -139,19 +141,32 @@ void SpriteSet::drawPrimitive(RenderContext* renderContext, int index)
   Matrix4x4 *modelMatrix = &renderContext->subscene->modelMatrix;
   
   if (fixedSize) {
-    float winwidth  = (float) renderContext->rect.width;
-    float winheight = (float) renderContext->rect.height;
+    Subscene* subscene = renderContext->subscene;
+    float winwidth  = (float) subscene->pviewport.width;
+    float winheight = (float) subscene->pviewport.height;
     // The magic number 27 is chosen so that plotmath3d matches text3d.
     float scalex = 27.0f/winwidth, scaley = 27.0f/winheight;
-    v3 =  (p * m) * o;
-    *modelMatrix = Matrix4x4::translationMatrix(v3.x, v3.y, v3.z)*
+    if (!rotating) {
+      v3 =  p * (m * o);
+      *modelMatrix = Matrix4x4::translationMatrix(v3.x, v3.y, v3.z)*
                    Matrix4x4::scaleMatrix(scalex, scaley, (scalex + scaley)/2.0f);
+    } else {
+      UserViewpoint* userviewpoint = subscene->getUserViewpoint();
+      /* FIXME: The magic value below is supposed to approximate the non-rotating size. */
+      float zoom = userviewpoint->getZoom(),
+        scale = zoom * sqrt(scalex * scaley) * 4.0f;
+      *modelMatrix = m * Matrix4x4::translationMatrix(o.x, o.y, o.z)*
+        Matrix4x4::scaleMatrix(scale, scale, scale);
+    }
   } else {
     s = s * 0.5f;	
-    v3 = m * o;
-    *modelMatrix = Matrix4x4::translationMatrix(v3.x, v3.y, v3.z);
+    if (!rotating) {
+      v3 = m * o;
+      *modelMatrix = Matrix4x4::translationMatrix(v3.x, v3.y, v3.z);
+    } else
+      *modelMatrix = m * Matrix4x4::translationMatrix(o.x, o.y, o.z);
   }
-  
+    
   if (pos.size())
     getAdj(index);
   
@@ -286,7 +301,7 @@ int SpriteSet::getAttributeCount(SceneNode* subscene, AttribID attrib)
       if (!shapes.size()) return 0;
       else return 4;
     }
-    case FLAGS:	   return 2;
+    case FLAGS:	   return 3;
     case ADJ: return 1;
     case POS: return pos.size();
   }
@@ -331,8 +346,9 @@ void SpriteSet::getAttribute(SceneNode* subscene, AttribID attrib, int first, in
         }
         return;
       case FLAGS:
-      	if (first == 0) *result++ = (double) ignoreExtent;
-      	*result++ = (double) fixedSize;
+      	if (first < 1) *result++ = (double) ignoreExtent;
+      	if (first < 2 && n > 1) *result++ = (double) fixedSize;
+      	if (n > 2) *result++ = (double) rotating;
       	return;
       case ADJ:
         if (pos.size() > 0) {